Updated seeding scripts
