+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "seed": "node database/seed.js",
-    "seed:s3": "node database/seed-s3.js",
+    "seed:s3": "node deploy/upload-photos.js",
     "start": "npx webpack --watch --mode development",
     "build": "npx webpack --mode development",
     "server": "node server/index.js",
